Always map Eisu on Apple Keyboard to Muhenkan with AutoHotkey
This was intended to work around an issue where the ReadLine of
PowerShell reads `Muhenkan` as `@`. Now that there is such a setting in
Windows Terminal to make `vk(29)` (Muhenkan) harmless by mapping it to
`Terminal.ResetFontSize` and `conhost.exe` is hardly used nowadays, I am
removing this work around for now.